fix(delete)quick-fix

+++ commands/delete.js
@@ -1,14 +1,11 @@
 exports.run = (client, message, args) => {
-  if (!args[0]) {
-    return message.channel.send("Please specify a number of messages to delete.");
+  if (!args[0] || args[0] > 99 || args[0] < 1) {
+    return message.channel.send("Please specify a number of messages to delete, between 1 and 99");
   }
   let toDelete = parseInt(args[0]) + 1;
   message.channel.send(`Deleting ${toDelete} messages.`)
-    .then(message => {
-      message.channel.bulkDelete(toDelete)
-        .then(async () => {
-          await message.delete();
-        })
+    .then(async message => {
+      await message.channel.bulkDelete(toDelete)
     })
 };
